Create spec of all packages to be installed by all jobs
In normal bosh, a job would depend on a subset of packages
During agent #job_install, only those subset of packages would be installed
and any other already-installed packages would be removed

Because of the removal step, we will tell the agent to install
all packages for all jobs.

As there is no blobstore, we stub the HTTP request to the blobstore
to return the compiled_package blob; using webmock's #stub_request
ckages_spec = {}
For each package, add the latest compiled_package; e.g.
{
  "common" => { "name" => "common", "version" => "0.2-dev", "sha1" => Digest::SHA1.file(package_common_tgz).hexdigest, "blobstore_id" => "package_common_blobstore" }
}